refactor: flatten the plugin into the composepwa package
The tasks and tasks.shared packages date from when the plugin defined six task
classes; since the staging moved into InitComposePwa they held one task and two
files of plain functions. Everything now lives flat in dev.yuyuyuyuyu.composepwa,
matching the plugin id.

Breaking (alpha): the task class is now
dev.yuyuyuyuyu.composepwa.InitComposePwa. Applying the plugin by id is
unaffected.
