+++ plugin/build.gradle.kts
@@ -35,7 +35,7 @@ gradlePlugin {
             description =
                 "This Gradle plugin provides a Gradle Task, that build your Compose Multiplatform Web App as a PWA."
             tags = listOf("compose multiplatform", "web", "pwa")
-            implementationClass = "dev.yuyuyuyuyu.ComposePwa"
+            implementationClass = "dev.yuyuyuyuyu.composepwa.ComposePwa"
         }
     }
 }
