+++ plugin/src/main/kotlin/dev/yuyuyuyuyu/composepwa/ComposePwa.kt
@@ -1,7 +1,6 @@
-package dev.yuyuyuyuyu
+package dev.yuyuyuyuyu.composepwa
 
 import com.github.gradle.node.npm.task.NpxTask
-import dev.yuyuyuyuyu.tasks.InitComposePwa
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
